+++ src/DocSharp.Docx/DocxToMarkdownConverter.cs
@@ -36,7 +36,9 @@ public class DocxToMarkdownConverter : DocxToStringWriterBase<MarkdownStringWrit
 
     /// <summary>
     /// By default only inline images are supported,  
-    /// because other DOCX image layouts have no direct equivalent in HTML/Markdown and can lead to unexpected results.  
+    /// because other DOCX image layouts have no direct equivalent in HTML/Markdown  
+    /// (or important properties are missing such as "both sides" for square layout)  
+    /// and can lead to unexpected results.  
     /// However, if desired, this property can be set to ImageLayoutType.InlineAndAnchored 
     /// to preserve the "top and bottom", "square", "tight" and "through" wrap layouts too, 
     /// or to ImageLayoutType.All to preserve absolutely positioned images ("in front of"/"behind" text) too.
@@ -723,9 +725,9 @@ public class DocxToMarkdownConverter : DocxToStringWriterBase<MarkdownStringWrit
             return;
         }
 
-        if (!string.IsNullOrWhiteSpace(ImagesOutputFolder))
+        if (!string.IsNullOrWhiteSpace(ImagesOutputFolder) && 
+            VmlHelpers.IsLayoutSupported(element, SupportedImagesLayout))
         {
-            // TODO: detect inline / anchored / floating and hyperlink for VML images
             if (element.GetFirstDescendant<V.ImageData>() is V.ImageData imageData &&
                 imageData.RelationshipId?.Value is string relId)
             {